pyproject.toml
[tool.poetry]
name = "mlflow-hydra-exp"
version = "0.1.0"
description = ""
authors = ["Andre1M <andrei95@gmail.com>"]
readme = "README.md"
packages = [{include = "mlflow_hydra_exp"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.10"
numpy = "^1.23.2"
torch = "^1.12.1"
torchvision = "^0.13.1"
catalyst = {extras = ["ml"], version = "^22.4"}
tensorboard = "^2.10.0"
omegaconf = "^2.2.3"
hydra-core = "^1.2.0"
mlflow = "^1.28.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
